Implement context for state and constants for action suite
