+++ frontend/src/constants/ACTIONS.js
@@ -1,0 +1,9 @@
+const ACTIONS = {
+  SET_CURRENT_PLAY_AREA: 'SET_CURRENT_PLAY_AREA',
+  OPEN_QR: 'OPEN_QR',
+  CLOSE_MODAL: 'CLOSE_MODAL',
+  // API Actions
+  UPDATE_USER_DATA: 'UPDATE_USER_DATA',
+};
+
+export { ACTIONS };
