+++ frontend/src/contexts/userContext.js
@@ -1,0 +1,4 @@
+import React from 'react';
+
+const UserContext = React.createContext();
+export { UserContext };
